MDL-59911 fileconverter_unoconv: Fix: test PDF fails after upgrade
When generating the PDF of unoconv_test.docx this file's previous conversions
are removed from mdl_file_conversion.  But these may also be removed by
\core_files\task\conversion_cleanup_task in which case an error occurs as
delete() is invoked with a NULL id.  Added check for non-NULL id prior to
invoking delete().

+++ files/converter/unoconv/classes/converter.php
@@ -194,7 +194,9 @@ class converter implements \core_files\converter_interface {
 
         $conversions = conversion::get_conversions_for_file($testdocx, $format);
         foreach ($conversions as $conversion) {
-            $conversion->delete();
+            if ($conversion->get('id')) {
+                $conversion->delete();
+            }
         }
 
         $conversion = new conversion(0, (object) [
